+++ cardano-cli/src/Cardano/CLI/Ping.hs
@@ -56,18 +56,20 @@ data PingCmd = PingCmd
   , pingCmdMagic    :: !Word32
   , pingCmdJson     :: !Bool
   , pingCmdQuiet    :: !Bool
+  , pingCmdQuery    :: !Bool
   } deriving (Eq, Show)
 
 pingClient :: Tracer IO CNP.LogMsg -> Tracer IO String -> PingCmd -> [CNP.NodeVersion] -> AddrInfo -> IO ()
 pingClient stdout stderr cmd = CNP.pingClient stdout stderr opts
   where opts = CNP.PingOpts
-          { CNP.pingOptsQuiet     = pingCmdQuiet cmd
-          , CNP.pingOptsJson      = pingCmdJson cmd
-          , CNP.pingOptsCount     = pingCmdCount cmd
-          , CNP.pingOptsHost      = maybeHostEndPoint (pingCmdEndPoint cmd)
-          , CNP.pingOptsUnixSock  = maybeUnixSockEndPoint (pingCmdEndPoint cmd)
-          , CNP.pingOptsPort      = pingCmdPort cmd
-          , CNP.pingOptsMagic     = pingCmdMagic cmd
+          { CNP.pingOptsQuiet           = pingCmdQuiet cmd
+          , CNP.pingOptsJson            = pingCmdJson cmd
+          , CNP.pingOptsCount           = pingCmdCount cmd
+          , CNP.pingOptsHost            = maybeHostEndPoint (pingCmdEndPoint cmd)
+          , CNP.pingOptsUnixSock        = maybeUnixSockEndPoint (pingCmdEndPoint cmd)
+          , CNP.pingOptsPort            = pingCmdPort cmd
+          , CNP.pingOptsMagic           = pingCmdMagic cmd
+          , CNP.pingOptsHandshakeQuery  = pingCmdQuery cmd
           }
 
 runPingCmd :: PingCmd -> ExceptT PingClientCmdError IO ()
@@ -89,7 +91,7 @@ runPingCmd options = do
       return ([addr], CNP.supportedNodeToClientVersions $ pingCmdMagic options)
 
   -- Logger async thread handle
-  laid <- liftIO . async $ CNP.logger msgQueue $ pingCmdJson options
+  laid <- liftIO . async $ CNP.logger msgQueue (pingCmdJson options) (pingCmdQuery options)
   -- Ping client thread handles
   caids <- forM addresses $ liftIO . async . pingClient (Tracer $ doLog msgQueue) (Tracer doErrLog) options versions
   res <- L.zip addresses <$> mapM (liftIO . waitCatch) caids
@@ -194,5 +196,11 @@ pPing = PingCmd
         [ Opt.long "quiet"
         , Opt.short 'q'
         , Opt.help "Quiet flag, CSV/JSON only output"
+        ]
+      )
+  <*> ( Opt.switch $ mconcat
+        [ Opt.long "query"
+        , Opt.short 'q'
+        , Opt.help "Query flag."
         ]
       )